# backend/openapi.yaml
openapi: 3.1.0
info:
  title: LabWebSystem Backend API
  version: 1.0.0
  description: LWSのアプリ管理API
servers:
  - url: /api/v1
paths:
...
  /applications:
    get:
      operationId: listApplications
      responses:
        '200': { description: アプリ一覧, content: { application/json: { schema: { $ref: '#/components/schemas/ApplicationList' } } } }
    post:
      operationId: createApplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequest'
      responses:
        '202': { description: Operationを作成, content: { application/json: { schema: { $ref: '#/components/schemas/OperationReference' } } } }
  /applications/{application}:
    parameters: [{ name: application, in: path, required: true, schema: { type: string } }]
    get: { operationId: getApplication, responses: { '200': { description: アプリ取得, content: { application/json: { schema: { $ref: '#/components/schemas/Application' } } } }, '404': { description: 未検出 } } }
    patch:
      operationId: updateApplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateApplicationRequest'
      responses: { '202': { description: Operationを作成, content: { application/json: { schema: { $ref: '#/components/schemas/OperationReference' } } } } }
    delete:
      operationId: unregisterApplication
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActionRequest' }
      responses: { '202': { description: Operationを作成, content: { application/json: { schema: { $ref: '#/components/schemas/OperationReference' } } } } }
  /applications/{application}:start:
    parameters: [{ name: application, in: path, required: true, schema: { type: string } }]
    post:
      operationId: startApplication
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ActionRequest' } } } }
      responses: { '202': { description: Operationを作成, content: { application/json: { schema: { $ref: '#/components/schemas/OperationReference' } } } } }
  /applications/{application}:register:
    parameters: [{ name: application, in: path, required: true, schema: { type: string } }]
    post:
      operationId: registerApplication
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ActionRequest' } } } }
      responses: { '202': { description: Operationを作成, content: { application/json: { schema: { $ref: '#/components/schemas/OperationReference' } } } } }
  /applications/{application}:stop:
    parameters: [{ name: application, in: path, required: true, schema: { type: string } }]
    post:
      operationId: stopApplication
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ActionRequest' } } } }
      responses: { '202': { description: Operationを作成, content: { application/json: { schema: { $ref: '#/components/schemas/OperationReference' } } } } }
  /applications/{application}:sync:
    parameters: [{ name: application, in: path, required: true, schema: { type: string } }]
    post:
      operationId: syncApplication
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ActionRequest' } } } }
      responses: { '202': { description: Operationを作成, content: { application/json: { schema: { $ref: '#/components/schemas/OperationReference' } } } } }
  /applications/{application}:rebuild:
    parameters: [{ name: application, in: path, required: true, schema: { type: string } }]
    post:
      operationId: rebuildApplication
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ActionRequest' } } } }
      responses: { '202': { description: Operationを作成, content: { application/json: { schema: { $ref: '#/components/schemas/OperationReference' } } } } }
  /applications/{application}:purge:
    parameters: [{ name: application, in: path, required: true, schema: { type: string } }]
    post:
      operationId: purgeApplication
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PurgeRequest' } } } }
      responses: { '202': { description: Operationを作成, content: { application/json: { schema: { $ref: '#/components/schemas/OperationReference' } } } } }
  /applications/{application}/configuration:
    parameters: [{ name: application, in: path, required: true, schema: { type: string } }]
    get: { operationId: getConfiguration, responses: { '200': { description: 設定取得, content: { application/json: { schema: { $ref: '#/components/schemas/ConfigurationResponse' } } } }, '404': { description: 未検出 } } }
    patch:
      operationId: updateConfiguration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigurationRequest'
      responses: { '202': { description: Operationを作成, content: { application/json: { schema: { $ref: '#/components/schemas/OperationReference' } } } } }
  /operations/{operation}:
    parameters: [{ name: operation, in: path, required: true, schema: { type: string } }]
    get: { operationId: getOperation, responses: { '200': { description: Operation取得, content: { application/json: { schema: { $ref: '#/components/schemas/OperationResource' } } } } } }
  /operations/{operation}:watch:
    parameters: [{ name: operation, in: path, required: true, schema: { type: string } }]
    get: { operationId: watchOperation, responses: { '200': { description: SSE, content: { text/event-stream: { schema: { type: string } } } } } }
  /applications/{application}/logEntries:
    parameters:
      - { name: application, in: path, required: true, schema: { type: string } }
      - { name: view, in: query, required: true, schema: { type: string, enum: [task, application, related] } }
      - { name: service, in: query, schema: { type: string, pattern: '^[a-z0-9][a-z0-9_-]{0,127}$' } }
      - { name: startAt, in: query, schema: { type: string, format: date-time } }
      - { name: endAt, in: query, schema: { type: string, format: date-time } }
      - { name: cursor, in: query, schema: { type: string, minLength: 1 } }
      - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500 } }
    get:
      operationId: listLogEntries
      responses:
        '200': { description: ログ一覧, content: { application/json: { schema: { $ref: '#/components/schemas/LogEntryList' } } } }
        '400': { description: 検索条件が不正 }
        '404': { description: 未検出 }
  /applications/{application}/logEntries:watch:
    parameters:
      - { name: application, in: path, required: true, schema: { type: string } }
      - { name: view, in: query, required: true, schema: { type: string, enum: [task, application, related] } }
      - { name: service, in: query, schema: { type: string, pattern: '^[a-z0-9][a-z0-9_-]{0,127}$' } }
      - { name: after, in: query, schema: { type: string, minLength: 1 } }
    get:
      operationId: watchLogEntries
      responses:
        '200': { description: SSE, content: { text/event-stream: { schema: { type: string } } } }
        '400': { description: 検索条件が不正 }
        '404': { description: 未検出 }
components:
  schemas:
    OperationReference:
      type: object
      required: [name]
      properties:
        name: { type: string, pattern: '^operations/.+$' }
    OperationResource:
      type: object
      required: [name, state]
      properties:
        name: { type: string, pattern: '^operations/.+$' }
        state: { type: string, enum: [queued, running, succeeded, failed, cancelled] }
        errorMessage: { type: string }
    Application:
      type: object
      required: [name, subdomain, repositoryUrl, ref, desiredState, observedState, registrationState, observedAt, reconciling, etag]
      properties:
        name: { type: string, pattern: '^applications/.+$' }
        subdomain: { type: string }
        repositoryUrl: { type: string }
        ref: { type: string }
        desiredState: { type: string }
        observedState: { type: string }
        registrationState: { type: string, enum: [ACTIVE, UNREGISTERED] }
        observedAt: { type: string, format: date-time }
        reconciling: { type: boolean }
        latestOperation: { type: string }
        etag: { type: string }
    ApplicationList:
      type: object
      required: [applications]
      properties:
        applications: { type: array, items: { $ref: '#/components/schemas/Application' } }
    LogEntry:
      type: object
      required: [id, cursor, occurredAt, level, component, message]
      properties:
        id: { type: string }
        cursor: { type: string }
        occurredAt: { type: string, format: date-time }
        level: { type: string, enum: [debug, info, warn, error] }
        component: { type: string, enum: [backend, caddy, coredns, dashboard, application] }
        applicationId: { type: string }
        operationId: { type: string }
        service: { type: string }
        containerName: { type: string }
        message: { type: string }
    LogEntryList:
      type: object
      required: [entries, liveCursor]
      properties:
        entries: { type: array, items: { $ref: '#/components/schemas/LogEntry' } }
        nextCursor: { type: string }
        liveCursor: { type: string }
    ConfigurationResponse:
      type: object
      required: [variables]
      properties:
        variables: { type: array, items: { type: object, required: [name, isSecret], properties: { name: { type: string }, isSecret: { type: boolean } } } }
    CreateApplicationRequest:
      type: object
      required: [repositoryUrl, ref, subdomain, requestId]
      properties:
        repositoryUrl: { type: string, format: uri }
        ref: { type: string, minLength: 1 }
        subdomain: { type: string, pattern: '^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$' }
        requestId: { type: string, format: uuid }
    UpdateApplicationRequest:
      type: object
      required: [requestId]
      properties:
        ref: { type: string }
        subdomain: { type: string }
        requestId: { type: string, format: uuid }
    ConfigurationRequest:
      type: object
      required: [variables, requestId]
      properties:
        variables: { type: object, additionalProperties: { $ref: '#/components/schemas/Variable' } }
        requestId: { type: string, format: uuid }
    ActionRequest:
      type: object
      required: [requestId]
      properties:
        requestId: { type: string, format: uuid }
    PurgeRequest:
      type: object
      required: [requestId, confirm]
      properties:
        requestId: { type: string, format: uuid }
        confirm: { type: boolean, const: true }
    Variable:
      type: object
      required: [value, secret]
      properties:
        value: { type: string }
        secret: { type: boolean }
...